+++ src/content-script/ui-tweaks.ts
@@ -274,6 +274,8 @@ function alternateMessageQuery({
   $query.append($newQueryWrapper);
 
   $query.off('dblclick').on('dblclick', () => {
+    if ($query.find('textarea').length) return;
+    
     const $buttonBar = $messageBlock.find(
       '.mt-sm.flex.items-center.justify-between'
     );